Change foreach macro used in q_free
In the previous version of q_free, list_for_each_safe is used to
iterate through the queue. Each node in the queue is then removed
at each iteration.

As each allocated space would be free right after list_del,
Removing the nodes before releasing the element would make no
difference compared to just releasing the elements. Hence,
list_del is removed.

Both list_for_each_safe and list_for_each_entry_safe had the same
condition for the for loop, which they would stop iterating when
the node of current iteration is equal to head of the list.
When the queue is empty, from q_new we know that the next pointer
of the queue points to itself, so the list would not even enter
the foreach loop. Thus, list_empty is not needed at all.