Complete implementation of q_insert
Implementation of q_insert_head and q_insert_tail is implemented
using functions / macros included in list.h as well as strncpy from
string.h, the program can now insert string to head or tail of the
queue.

Both implementation is tested, with ih and it command using the
interpreter qtest. Though there is a lot of room of optimization,
this should be enough for a placeholder to run all test case.