Complete implementation of q_remove
Implementation of q_remove_head and q_remove_tail is implemented
using functions / macros included in list.h as well as strncpy from
string.h. The program can now remove the first and last element from
the queue, the value of the removed queue would be copied to the
memory indicated by the second argument provided the argument is
not NULL.

Both implementation is tested, with rh and rt after inserting to
queue with ih and it commands using the interpreter qtest.
Both function is a placeholder to run all test case and would be
revisited later for further optimization.