Fix bug q_sort not linking prev chain properly
In the previous version of q_sort, q_sort only link the first and
last element of queue to head of list at the end of function.

As merge_sort sort the list as if it is a singly-linked list,
merge_sort would not maintain the prev link in each struct list_head
structure.

qtest would not show this error as the checks done in show_queue()
are is_circular, which returns true when the pointer moving along
the prev chain meets the pointer moving along the next chain.
Which will not report error when prev is not correctly linked.

The print process also do not show a incorrect result as it only
print by moving along the next chain, hence the prev chain is
ignored.

With this bug fixed, make test can now pass both test case
trace-05-ops and trace-06-ops.