Fix bug remove element not considering buffer size
In the previous version of both q_remove_head and q_remove_tail,
the program does not take into consideration of bufsize of sp.

Even though the program copies only $bufsize characters to sp, which
prevents buffer overflow from occuring, when the length of value in
the element to be removed is larger than bufsize, sp is not NULL
terminated.

When the sp is displayed to the console, as it is not NULL terminated,
the program may still access memory that is not part of sp. Which
results in error.

This bug is detected by the test case trace-07-string, with this bug
fixed, make test can now pass the test case trace-07-string