Optimize merge sort when splitting list
In the previous version of merge sort, the recursive function splits
the list in a ratio of 1 : (n-1), which causes the merging process
have time complexity of O(n^2).

The merge sort is updated to splits the list in a ratio of 1:1,
using Floyd's tortoise and hare cycle detection algorithm. Which
conviniently have the tortoise stops at about the middle of list
when the hare reaches the end of list.

The algorithm now have time complexity of O(nlogn) which is shown
by passing trace-15-perf