Make valgrind stop complaining
In previous version of program, when deleting a node, even though
we does not do anything on the pointer itself, the pointer will
move to a position where it is illegal and valgrind does not like
that. Now, the pointer won't continue to move as long as it pass
its boundary.